Multivariate Analysis Mideterm Project
"This project is an assignment for the course Multivariate Analysis, conducted during the first semester of the academic year 2024. It encompasses content covered in the course from the beginning of the term up to midterm. The objective of this project is to apply the knowledge gained from the coursework in developing a tool independently."